+++ geoapps/inversion/potential_fields/gravity/constants.py
@@ -575,7 +575,7 @@ validations = {
 validations = dict(base_validations, **validations)
 
 app_initializer = {
-    "geoh5": "../../../assets/FlinFlon.geoh5",
+    "geoh5": "../../../../assets/FlinFlon.geoh5",
     "forward_only": False,
     "data_object": UUID("{538a7eb1-2218-4bec-98cc-0a759aa0ef4f}"),
     "gz_channel": UUID("{6de9177a-8277-4e17-b76c-2b8b05dcf23c}"),
